+++ packages/ripple-binary-codec/src/types/index.ts
@@ -11,6 +11,7 @@ import { Currency } from './currency'
 import { Hash128 } from './hash-128'
 import { Hash160 } from './hash-160'
 import { Hash256 } from './hash-256'
+import { IssuedCurrency } from './issued-currency'
 import { PathSet } from './path-set'
 import { STArray } from './st-array'
 import { STObject } from './st-object'
@@ -19,6 +20,8 @@ import { UInt32 } from './uint-32'
 import { UInt64 } from './uint-64'
 import { UInt8 } from './uint-8'
 import { Vector256 } from './vector-256'
+import { XChainAttestationBatch } from './xchain-attestation-batch'
+import { XChainBridge } from './xchain-bridge'
 
 const coreTypes = {
   AccountID,
@@ -28,6 +31,7 @@ const coreTypes = {
   Hash128,
   Hash160,
   Hash256,
+  IssuedCurrency,
   PathSet,
   STArray,
   STObject,
@@ -36,6 +40,8 @@ const coreTypes = {
   UInt32,
   UInt64,
   Vector256,
+  XChainAttestationBatch,
+  XChainBridge,
 }
 
 Object.values(Field).forEach((field) => {