+++ packages/xrpl/HISTORY.md
@@ -1,11 +1,13 @@
 # xrpl.js (ripple-lib) Release History
 
 Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xrpl-announce) for release announcements. We recommend that xrpl.js (ripple-lib) users stay up-to-date with the latest stable release.
+
 ## Unreleased
+### Added
+* Support for the cross-chain bridge feature.
 
 ## 2.4.0 (2022-09-01)
 ### Added
-
 - Export `verify` from ripple-keypairs as `verifyKeypairSignature` for use in web-apps.
 
 ### Fixed